Make test suite more robust
The test script had special handling for input test files with a
shebang. What normally is the interpreted is treated as arguments to
indent-sort to produce the desired .correct file. However, we didn't
have a way to distinguish this from a normal shebang. So now we'll put a
space for the test scripts.